Refactor HashAggregationOperator
Extracted interface for HashAggregationBuilder with separate implementation for spillable version and in memory version.
